Refactor implementation for clarity and usage
Two changes:
    1. Prefer explicit implementation of ISerialize on user types
    2. Use wrapper structs for built-ins and extension methods

(1) should improve the experience by hiding the Serialize method on user
types, which usually isn't very helpful. The intended usage is for users
to call some method on one of the serializer implementations, like
JsonSerializer, not to call Serialize manually.

(2) changes how we deal with provided implementations for core framework
types.  Previously, overloads were provided on each specialty interface
(like ISerializeType) for each of the built-in types. This works but
forces the implementor of a serializer to have a lot of boilerplate when
the correct implementation is almost certainly to call the
ISerialize.Serialize(<core-type>) overload. Now we provide a wrapper
type that implements the interface for the core types and have the
generator generate the creation.