Remove generics from ISerialize (#22)
There are serious limitations in the readability of the generics used in
the existing Serde model. Using boxed interfaces and dynamic dispatch is
clearer, at the potential expense of performance.

The previous implementation has been moved to a new interface hierarchy, ISerializeStatic.
Source generator support for ISerializeStatic can be added back later.