Add support for constructors (#75)
The IDeserialize generator currently always expects an accessible parameterless construct and that the data members (public fields and properties) are all writable.

This change adds support for specifying a particular constructor to call during serialization. It's expected that the constructor parameter names match the names of the data members exactly (no casing differences). Any remaining data members are still expected to exist, be writable, and are set in the type initializer after calling the constructor.