Add XML serialization support
Adds very simple support for XML as a backend (serialization only). This
required a new feature, attribute passthrough. This feature allows the
frontend source generator to gather a list of all the attributes on a
field and pass them to the backend. This feature is not enabled by
default, so if a user wants to use this feature to configure their XML
serialization they will need to specify it in SerdeMemberOptions (soon
to be added to SerdeTypeOptions as well).

A number of other small changes were also necessary, including adding
support for System.Decimal, System.Double, and System.Single. Wrappers
have also been added for those types.