Change default field format to camelCase (#89)
* Change default formatting to camelCase

The current default formatting is to write fields with exactly the
same casing as they are in the source code. However, the vast majority
of JSON files are formatted with camelCase. Since JSON is the most popular
format, this means that the most common serialization format is camelCase.
Thus, it makes sense for Serde to write camelCase by default, especially
since it incurs no performance cost due to source generation.

To achieve the old behavior, users can pass MemberFormat.None.