Add support for nested explicit wrappers (#140)
Sometimes you need to fully specify a wrapper type. This is most common
for situations when you need a custom wrapper type, but that type is
embedded inside an outer generic type (like a List) that needs its own
wrapper. In this case, it's not enough to specify the outer wrapper (the
List wrapper) or the inner wrapper (the custom wrapper) alone. You need
the full composition.

The new `WrapperSerialize` and `WrapperDeserialize` properties on the
GenerateSerialize, GenerateDeserialize, and GenerateSerde attributes can
be used for this.