Add FieldMap abstraction (#162)
FieldMap is a simple way of mapping field names to indices, which can be used
for deserialization. The dispatch path looks similar to the existing visitor
pattern, but results in less code both on the generation side and on the compiler
side. It's one component in the work-in-progress in minimizing generic interface
dispatch.