Provide limited support for serializing unions
Since discriminated unions have yet to ship as an actual language
feature, this change supports a workaround used by StaticCS to encode a
discriminated union. In fact, JsonValue is one such DU, but it uses an
"untagged" representation, which is not currently auto-generated.

The default representation is the same one used by serde.rs, which is
referred to as "externally tagged." Other encodings are possible, but
not yet implemented.