Remove visitor pattern from IDeserialize
This is a very different approach from the rest of the library and only
really addresses a single scenario -- deserializing an arbitrary value
with no type information. That's not a core scenario so we should skip
it and come back later if there's a good way to fit it into the current
model.