Change enum proxy deserialize (#250)
The existing implementation expects to be able to decode an enum from
TryReadIndex. This means that it can only rely on information from the
ISerdeInfo. ISerdeInfo only has access to the name, not the enum value,
which means it's not possible for deserializer implementations to use
the enum value without a custom proxy. By checking for EndOfType,
deserializer implementations can indicate that they want to encode the
value directly.